requirements_soa.txt
spyne==2.13.16
Flask==2.3.3
lxml==4.9.3
requests==2.31.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9